feat: decimal weights, paused-by-default sessions, compact UI
- Task weight is now Double. Input accepts comma or dot (1,5 / 1.5).
  Display formats integers as "x2", decimals as "x1.5".
- DB to v3 with real migrations (v1->v2 added description/enabled,
  v2->v3 widens weight to REAL). Destructive migration kept as
  fallback only; existing tasks are preserved on upgrade.
- Work screen starts paused. Pressing the primary Play/Pause button
  begins/resumes the timer. When a task auto-completes or is skipped,
  the next-most-behind enabled task is loaded already paused so the
  user controls when the next session starts.
- Work screen layout: primary Play/Pause big at the bottom, Saltar
  and Detener as outlined secondaries above it. Label switches to
  "Empezar" when elapsed is 0, "Reanudar" mid-session.
- Home screen: FAB replaced by a tonal "+" button paired in the same
  row as "Iniciar trabajo" so both controls share one baseline.
- Compact task rows: description truncated to one line; time and
  edit/delete icons share the same row to save vertical space.

+++ app/src/main/java/com/rotask/data/Task.kt
@@ -8,7 +8,7 @@ data class Task(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
     val description: String = "",
-    val weight: Int,
+    val weight: Double,
     val enabled: Boolean = true,
     val debtSeconds: Long = 0L
 )
